%Asservissement_HL : correction de la variable de conssigne d'angle %Asservissement_LL : correction du calcul des consignes moteur +const.h ajout d'une constante d'entraxe diamètre

+++ Headers/Asservissement_HL.h
@@ -7,7 +7,7 @@
 void Mode_Asserv(int mode);
 int get_Mode_Asserv();
 extern signed long Asserv_Cons_distance;
-extern signed long Asserv_Cons_angle;
+extern long double Asserv_Cons_angle;
 void Gestion_Asserv_HL(long Tick_droit,long Tick_gauche,long* ordre_distance,long* ordre_angle);
 void Asserv_Reset_Integrateur(void);
 void Asserv_Reset_Derivateur(void);
